chore: change `RecvWithContext` signature
Nobody likes current `RecvWithContext` implementation because you don't really know what happens inside.
Did it return false because channel was closed? Did it return because context was canceled?

The problem manifested itself with this pattern being repeated in Talos code:

```
val, ok := channel.RecvWithContext(ctx, someCh);
if !ok && ctx.Err() != nil {
	return nil //nolint:nilerr
}
```

This is hard to remember and even harder to write. I believe we can do better. By changing the result
from `bool` to `RecvState` we can now write code like this:

```
val, state := channel.RecvWithContext(ctx, someCh);
if state == channel.StateCancelled {
	return nil
}
```
Which is still node ideal for cases when we call it on `EventCh`, but that should be improved in the following PR to the runtime.

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>