chore: add hashtriemap implementation
This PR adds a hashtriemap implementation to the project. This is mostly a direct copy from unique package [^1]
internals [^2] with some adjustments to make it work with our project. Once `maphash.Comparable` lands we can remove
dependency on Go internals [^3]. Michael Knyszek says he plans to use this implementation as a general idea for the
future generic sync/v2.Map, but nothing is in stone yet.

[^1]: https://github.com/golang/go/issues/62483
[^2]: https://go-review.googlesource.com/c/go/+/573956
[^3]: https://github.com/golang/go/issues/54670

Signed-off-by: Dmitriy Matrenichev <dmitry.matrenichev@siderolabs.com>